algorithm optimization
When you calculate the expectation of the same colored K4, you don't need to iterate all possible K4. Because if you color one edge, only C(n,2) K4s are impacted. So just calculate the sum of C(n,2) K4s' expectation, the other K4s' expectation is the same whether you color the edge black or white.
I have made the experiment, the result shown that this way is faster than the original one and the accuracy is the same.